+++ src/main/java/com/power/doc/utils/DocUtil.java
@@ -582,11 +582,13 @@ public class DocUtil {
             case "localdatetime":
             case "java.time.localdatetime":
             case "java.time.localdate":
+            case "java.time.Period":
             case "localdate":
             case "offsetdatetime":
             case "localtime":
             case "timestamp":
             case "zoneddatetime":
+            case "period":
             case "java.time.zoneddatetime":
             case "java.time.offsetdatetime":
             case "java.lang.character":